feat(e2e-core): a fallback key, so an exhausted device stays reachable
Publishing only one-time keys meant a device whose supply ran out simply
could not be reached: the sender had no key to fetch, and nothing in the
flow said why. That is the hole X3DH's signed prekey exists to close, and
vodozemac has had the mechanism all along.

Every bundle now carries a fallback key. It is not consumed by use, so it
keeps a device reachable indefinitely - which is also the catch: many
senders may share one, so it gives weaker forward secrecy than a one-time
key. It is a fallback in the real sense, and the backend must keep serving
one-time keys whenever it still has them.

Rotation is the app's half of the deal. rotate_fallback_key() issues a new
one; vodozemac keeps the previous key alive so first messages already in
flight still open; forget_previous_fallback_key(), a rotation later, shuts
that window. Skip the second call and the first accomplishes nothing.

Also adds a cargo-audit config, which earned its place immediately. A
by-hand pass over the direct dependencies had reported a clean tree.
Across all 254 transitive crates there were ten advisories - most of them
high, nearly all in the libcrux stack that backs the post-quantum X-Wing
suite. Two were fixable and are fixed. The other nine are pinned out of
reach by hpke-rs, and the only route to them drags the whole MLS stack
onto release candidates, which is not a trade worth making for code that
protects real conversations. They are recorded in .cargo/audit.toml with
reasoning and a re-check trigger, as a liability register rather than a
mute button.

The CI workflow that runs this on every change is held back: the push
credential lacks GitHub's `workflow` scope, so it needs a token that has
it. Without that job, audit.toml is documentation rather than a gate.

93 tests passing, 4 soak runs, fmt and clippy clean.

+++ packages/e2e-core/tests/soak.rs
@@ -12,7 +12,10 @@ use std::time::Instant;
 use voiid_e2e_core::{api, GroupMember, Session};
 
 fn env_usize(key: &str, default: usize) -> usize {
-    std::env::var(key).ok().and_then(|v| v.parse().ok()).unwrap_or(default)
+    std::env::var(key)
+        .ok()
+        .and_then(|v| v.parse().ok())
+        .unwrap_or(default)
 }
 
 /// 1:1: sustained bidirectional traffic with periodic out-of-order delivery and
@@ -95,7 +98,10 @@ fn soak_group_churn() {
         for j in 0..msgs_per_round {
             let payload = format!("r{round}-m{j}");
             let ct = ag.encrypt(&alice, payload.as_bytes()).unwrap();
-            assert_eq!(bg.decrypt(&bob, &ct).unwrap().as_deref(), Some(payload.as_bytes()));
+            assert_eq!(
+                bg.decrypt(&bob, &ct).unwrap().as_deref(),
+                Some(payload.as_bytes())
+            );
             total_msgs += 1;
         }
 
@@ -139,7 +145,10 @@ fn soak_media_stream() {
         assert_eq!(dec, plaintext, "media #{i} ({size} bytes) round-trip");
     }
 
-    println!("soak media: {count} attachments in {:.1}s", start.elapsed().as_secs_f64());
+    println!(
+        "soak media: {count} attachments in {:.1}s",
+        start.elapsed().as_secs_f64()
+    );
 }
 
 /// Grow a group to N leaves and measure what that costs on the wire.
@@ -230,5 +239,5 @@ fn group_scale_welcome_and_tree_size() {
         last_tree as f64 / 1_048_576.0,
         commit_total as f64 / 1_048_576.0,
     );
-    assert_eq!(ag.member_count() as usize, members + 1);
+    assert_eq!(ag.member_count(), members + 1);
 }
